+++ go.mod
@@ -1,6 +1,6 @@
-module git.jwn.app/TM00458/Operations/scripts/ns-report-gen
+module github.com/JeremyPDonahue/custom_reporter
 
-go 1.18
+go 1.19
 
 require (
 	k8s.io/api v0.26.0
@@ -27,7 +27,6 @@ require (
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
-	github.com/sirupsen/logrus v1.9.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	golang.org/x/net v0.3.1-0.20221206200815-1e63c2f08a10 // indirect
 	golang.org/x/oauth2 v0.0.0-20220223155221-ee480838109b // indirect
